extract _monaco
